Move another hook
