+++ src/components/results/securityanalysis/security-analysis-result-tab.tsx
@@ -45,7 +45,7 @@ import { useParameterState } from 'components/dialogs/parameters/use-parameters-
 import { useNodeData } from 'components/use-node-data';
 import GlobalFilterSelector from '../../global-filter/global-filter-selector';
 import useGlobalFilters from '../common/global-filter/use-global-filters';
-import { useGlobalFilterOptions } from '../common/global-filter/use-global-filter-options';
+import { useGlobalFilterOptions } from '../../global-filter/use-global-filter-options';
 import { EQUIPMENT_TYPES } from '../../utils/equipment-types';
 import { usePaginationSelector } from 'hooks/use-pagination-selector';
